split monolith main.go into managable chunks
